% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfeed.R
\name{fetch_statfeed}
\alias{fetch_statfeed}
\title{Get the stat feed for a single match
Gets the stat feed for a single match}
\usage{
fetch_statfeed(matchId)
}
\arguments{
\item{matchId}{Match ID in format CD_M20262640101}
}
\value{
A dataframe
}
\description{
Get the stat feed for a single match
Gets the stat feed for a single match
}
\examples{
fetch_statfeed("CD_M20262640101")


}
