% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wExtendedStats.R
\name{fetch_stats_aflw}
\alias{fetch_stats_aflw}
\title{Get the extended player stats game by game
Gets the extended player stats game by game}
\usage{
fetch_stats_aflw(comp = "AFLW", season = NA, round = NA)
}
\arguments{
\item{comp}{= "AFLW", only services AFLW at the moment}

\item{season}{Season year, 2022 will return both 2022 seasons, will default to current year, can provide a vector e.g. c(2022:2026)}

\item{round}{Round numbers to return. NA is all rounds, can provide single round or multiple}
}
\value{
A dataframe
}
\description{
Get the extended player stats game by game
Gets the extended player stats game by game
}
\examples{
fetch_stats_aflw()
fetch_stats_aflw(round=c(2:4))


}
