% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i50r50.R
\name{get_next_i50_data}
\alias{get_next_i50_data}
\title{Produce a dataframe listing the i50 events and whether the next i50 was a repeat i50, score, or opponent i50
Produces a dataframe listing the i50 events and whether the next i50 was a repeat i50, score, or opponent i50}
\usage{
get_next_i50_data(statFeed)
}
\arguments{
\item{statFeed}{A statfeed as returned by fetch_statFeed}
}
\value{
A dataframe
}
\description{
Produce a dataframe listing the i50 events and whether the next i50 was a repeat i50, score, or opponent i50
Produces a dataframe listing the i50 events and whether the next i50 was a repeat i50, score, or opponent i50
}
